backend/requirements.txt
fastapi
uvicorn
pytesseract
pillow
pdfplumber==0.10.3
pdfminer.six==20221105
streamlit
requests